chore: rename script which adds file to gui repos
