+++ scripts/src/actions/fix-yaml-config.ts
@@ -1,5 +1,5 @@
 import 'reflect-metadata'
-import { addFileToGUIRepos } from './shared/add-gui-workflows'
+import { addFileToGUIRepos } from './shared/add-file-to-gui-repos'
 import {format} from './shared/format'
 
 addFileToGUIRepos()
